requirements.txt
PyYAML~=6.0
restocks-client~=0.1.2